Implemented randomised spawn
Make the player spawn at a random predefined position. To use drag and drop the SpawnPoint element to any place the player could spawn. One will be randomly picked on scene launch. If none are put the player will start where the player object is, and the visual markers will disappear.
The SpawnPoint GameObject is in the ExitSetup folder now renamed EntranceExitSetup.