Random shooter start
The shooter will now randomly pick a MainExit to spawn from. The code would not work if put in the ShooterAI file, so it is put in the ScenarioControl file to not create a new file (it seemed to me as the most logical choice).
Also a small insignificant minor change in the player spawn.